Enhancement: get ski by id  (#37)
* This prints HelloWorld using graphql now

Send a request and receive a response according to the fields you
specified.

* Updated Readme to show how to use Hello World

* Updated status.md

* Features two fields now for testing purposes + refactoring

The schema is now defined in a dedicated file located in a subdirectory
called "pplStock". There is a new field defined in the schema which needs
a parameter to generate its response.

* This is pretty basic. It only defines how to resolve the fields
which are requested. It does not access a database and it does not
define all necessary fields for a ski-set. It instead defines 3
basic fields: id, useCase, brand. The request needs a id as param,
it accesses a hardcoded database to get the response data.

* update readme and status